requirements/common.txt
# Core Stuff
# -------------------------------------
django==1.8.7
whitenoise==2.0.6

# Extensions
# -------------------------------------
Werkzeug==0.11.2
django-extensions==1.5.9
django-flat-theme==1.1.3
pytz==2015.7

# Configuration
# -------------------------------------
django-environ==0.4.0
django-sites==0.8
python-dotenv==0.3.0

# Models
# -------------------------------------
django-model-utils==2.4
django-uuid-upload-path==1.0.0
django-uuidfield==0.5.0

# Image Handling
# -------------------------------------
Pillow==3.0.0
django-versatileimagefield==1.2

# For the persistance stores
# -------------------------------------
psycopg2==2.6.1

# Django Rest Framework
# -------------------------------------
djangorestframework==3.3.1
